+++ Flutter.Backend.Service/Models/Requests/ConfirmOrderRequest.cs
@@ -10,7 +10,7 @@ namespace Flutter.Backend.Service.Models.Requests
 
         public string Description { get; set; }
 
-        public int Ship { get; set; }
+        public decimal Ship { get; set; }
 
         public Location AddressReceive { get; set; }
